pyproject.toml
[project]
name = "owl-ui-removal"
version = "0.1.0"
description = "UI mask prediction using Grounded-SAM-2 for video frame sequences"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "UI Tracker Team", email = "contact@example.com"}
]
license = {text = "MIT"}
keywords = ["ui", "segmentation", "tracking", "sam2", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "supervision>=0.16.0",
    "transformers>=4.35.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.64.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.0",
    "scipy>=1.9.0",
    "scikit-image>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/your-username/owl-ui-removal"
Documentation = "https://github.com/your-username/owl-ui-removal/blob/main/README.md"
Repository = "https://github.com/your-username/owl-ui-removal"
Issues = "https://github.com/your-username/owl-ui-removal/issues"

[project.scripts]
owl-ui-tracker = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.12.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[build-system]
requires = ["hatchling>=1.12.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ui_tracker", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
